Address CI concerns

+++ osu.Framework/MathUtils/Interpolation.cs
@@ -54,31 +54,31 @@ namespace osu.Framework.MathUtils
         }
 
         public static byte ValueAt(double time, byte val1, byte val2, double startTime, double endTime, EasingTypes easing = EasingTypes.None) =>
-            (byte)Math.Round(ValueAt(time, (double)val1, (double)val2, startTime, endTime, easing));
+            (byte)Math.Round(ValueAt(time, (double)val1, val2, startTime, endTime, easing));
 
         public static sbyte ValueAt(double time, sbyte val1, sbyte val2, double startTime, double endTime, EasingTypes easing = EasingTypes.None) =>
-            (sbyte)Math.Round(ValueAt(time, (double)val1, (double)val2, startTime, endTime, easing));
+            (sbyte)Math.Round(ValueAt(time, (double)val1, val2, startTime, endTime, easing));
 
         public static short ValueAt(double time, short val1, short val2, double startTime, double endTime, EasingTypes easing = EasingTypes.None) =>
-            (short)Math.Round(ValueAt(time, (double)val1, (double)val2, startTime, endTime, easing));
+            (short)Math.Round(ValueAt(time, (double)val1, val2, startTime, endTime, easing));
 
         public static ushort ValueAt(double time, ushort val1, ushort val2, double startTime, double endTime, EasingTypes easing = EasingTypes.None) =>
-            (ushort)Math.Round(ValueAt(time, (double)val1, (double)val2, startTime, endTime, easing));
+            (ushort)Math.Round(ValueAt(time, (double)val1, val2, startTime, endTime, easing));
 
         public static int ValueAt(double time, int val1, int val2, double startTime, double endTime, EasingTypes easing = EasingTypes.None) =>
-            (int)Math.Round(ValueAt(time, (double)val1, (double)val2, startTime, endTime, easing));
+            (int)Math.Round(ValueAt(time, (double)val1, val2, startTime, endTime, easing));
 
         public static uint ValueAt(double time, uint val1, uint val2, double startTime, double endTime, EasingTypes easing = EasingTypes.None) =>
-            (uint)Math.Round(ValueAt(time, (double)val1, (double)val2, startTime, endTime, easing));
+            (uint)Math.Round(ValueAt(time, (double)val1, val2, startTime, endTime, easing));
 
         public static long ValueAt(double time, long val1, long val2, double startTime, double endTime, EasingTypes easing = EasingTypes.None) =>
-            (long)Math.Round(ValueAt(time, (double)val1, (double)val2, startTime, endTime, easing));
+            (long)Math.Round(ValueAt(time, (double)val1, val2, startTime, endTime, easing));
 
         public static ulong ValueAt(double time, ulong val1, ulong val2, double startTime, double endTime, EasingTypes easing = EasingTypes.None) =>
-            (ulong)Math.Round(ValueAt(time, (double)val1, (double)val2, startTime, endTime, easing));
+            (ulong)Math.Round(ValueAt(time, (double)val1, val2, startTime, endTime, easing));
 
         public static float ValueAt(double time, float val1, float val2, double startTime, double endTime, EasingTypes easing = EasingTypes.None) =>
-            (float)ValueAt(time, (double)val1, (double)val2, startTime, endTime, easing);
+            (float)ValueAt(time, (double)val1, val2, startTime, endTime, easing);
 
         public static decimal ValueAt(double time, decimal val1, decimal val2, double startTime, double endTime, EasingTypes easing = EasingTypes.None) =>
             (decimal)ValueAt(time, (double)val1, (double)val2, startTime, endTime, easing);
